WIP: get rid of Band class
